feat: refactor code
Refactor code simplifying the main function and moving all Sudoku
functions to `src/solver.rs`.
